Add files via upload
These are the issue codes on GBIF as well as their extended names and explanations (although I've only added in the ones I found in the Geranium_data_clean file. I will continue adding in the other explanations as need be but so far the only ones I think may be worth considering are fpwktinv, rdativ, prenegion, conti, and prewcd.